requirements.txt
streamlit>=1.30.0
openai>=1.10.0
feedparser>=6.0.0
plotly>=5.18.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
